+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Neo4jAncestry Changelog
 
-* v0.0.1 - Initial Release with basic functionality.
+* v0.0.2 - Route traversing.
+ * TODO: `object.find_routes_to(other_object)`
+ * TODO: `object.find_routes_to(other_object, via: third_object)`
+ * TODO: `object.find_routes_to(other_object, via: [third_object, another_object])`
+  
+* v0.0.1 - Initial release with basic functionality.
  * Rake tasks for installation and setup of neo4j.
  * Basic traversing methods, such as `parents`, `children`, `ancestors`, `descendants`, `siblings`.